log validator: bookmarklet, more CORS: should now work fully everywhere
Turns out that if the validator page is served from https, then it can't
CORS-fetch files from http. Add protocol conversion with a warning.

Now it works both from rawgit and from people.mozilla.org (and locally),
regardless of the host or log URL protocols.

Also, the page now contains a bookmarklet which can be invoked while browsing
a talos log, and it will invoke the validator (only works when the validator is
hosted online).